==> 1.1.corruption.ipynb <==
<font color='lime'>

# Data corruption

</font> 

<font color='yellow'>

##  Load dataset

</font> 

import pandas as pd

data = pd.read_csv('data/spotify_songs.csv')

data.head()

<font color='yellow'>

##  Process the initial data

</font>

<font color='orange'>

- Corrupting 12% of the subset rows (introducing one empty cell)
- Saving the result

</font>

import numpy as np

# Calculate the number of rows to corrupt
num_rows_to_corrupt = int(0.12 * len(data))

# Randomly select rows to corrupt
rows_to_corrupt = np.random.choice(data.index, size=num_rows_to_corrupt, replace=False)

# For each selected row, randomly select a non-integer column to set to NaN
for row in rows_to_corrupt:
    col_to_corrupt = np.random.choice(data.columns)
    while data[col_to_corrupt].dtype == np.int64:
        col_to_corrupt = np.random.choice(data.columns)
    data.at[row, col_to_corrupt] = np.nan

# Save the corrupted dataframe to a new CSV file
corrupted_filename = "data/spotify_songs_corrupted.csv"
data.to_csv(corrupted_filename, index=False)

corrupted_filename
